+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/exec/ControllerImpl.java
@@ -90,6 +90,7 @@ import org.apache.druid.msq.indexing.MSQTuningConfig;
 import org.apache.druid.msq.indexing.MSQWorkerTaskLauncher;
 import org.apache.druid.msq.indexing.SegmentGeneratorFrameProcessorFactory;
 import org.apache.druid.msq.indexing.TaskReportMSQDestination;
+import org.apache.druid.msq.indexing.WorkerCount;
 import org.apache.druid.msq.indexing.error.CanceledFault;
 import org.apache.druid.msq.indexing.error.CannotParseExternalDataFault;
 import org.apache.druid.msq.indexing.error.FaultsExceededChecker;
@@ -1817,9 +1818,9 @@ public class ControllerImpl implements Controller
     int runningTasks = 1;
 
     if (taskLauncher != null) {
-      MSQWorkerTaskLauncher.WorkerStatus workerTaskStatus = taskLauncher.getWorkerTaskStatus();
-      pendingTasks = workerTaskStatus.getPendingWorkerCount();
-      runningTasks = workerTaskStatus.getRunningWorkerCount() + 1; // To account for controller.
+      WorkerCount workerTaskCount = taskLauncher.getWorkerTaskCount();
+      pendingTasks = workerTaskCount.getPendingWorkerCount();
+      runningTasks = workerTaskCount.getRunningWorkerCount() + 1; // To account for controller.
     }
     return new MSQStatusReport(
         taskState,
