+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/indexing/MSQWorkerTaskLauncher.java
@@ -349,12 +349,12 @@ public class MSQWorkerTaskLauncher
    * Returns a pair which contains the number of currently running worker tasks and the number of worker tasks that are
    * not yet fully started as left and right respectively.
    */
-  public WorkerStatus getWorkerTaskStatus()
+  public WorkerCount getWorkerTaskCount()
   {
     synchronized (taskIds) {
       int runningTasks = fullyStartedTasks.size();
       int pendingTasks = desiredTaskCount - runningTasks;
-      return new WorkerStatus(runningTasks, pendingTasks);
+      return new WorkerCount(runningTasks, pendingTasks);
     }
   }
 
@@ -468,32 +468,6 @@ public class MSQWorkerTaskLauncher
       if (Thread.interrupted()) {
         throw new InterruptedException();
       }
-    }
-  }
-
-  /**
-   * Information about current status of running tasks
-   */
-  public static class WorkerStatus
-  {
-    private final int runningWorkerCount;
-
-    private final int pendingWorkerCount;
-
-    public WorkerStatus(int runningWorkerCount, int pendingWorkerCount)
-    {
-      this.runningWorkerCount = runningWorkerCount;
-      this.pendingWorkerCount = pendingWorkerCount;
-    }
-
-    public int getRunningWorkerCount()
-    {
-      return runningWorkerCount;
-    }
-
-    public int getPendingWorkerCount()
-    {
-      return pendingWorkerCount;
     }
   }
