+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/indexing/WorkerCount.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.druid.msq.indexing;
+
+/**
+ * Information about current status of worker tasks
+ */
+public class WorkerCount
+{
+  private final int runningWorkerCount;
+
+  private final int pendingWorkerCount;
+
+  public WorkerCount(int runningWorkerCount, int pendingWorkerCount)
+  {
+    this.runningWorkerCount = runningWorkerCount;
+    this.pendingWorkerCount = pendingWorkerCount;
+  }
+
+  public int getRunningWorkerCount()
+  {
+    return runningWorkerCount;
+  }
+
+  public int getPendingWorkerCount()
+  {
+    return pendingWorkerCount;
+  }
+}
